pyproject.toml
[project]
name = "alapi-rag-starter"
version = "0.1.0"
description = "A ready-to-use FastAPI starter template for building RAG agents using alAPI"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "requests>=2.26.0",
    "pydantic>=1.8.0",
    "python-multipart>=0.0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
